build: enable type warnings, -O0 debug, fix warns
        This commit enables type warnings -Wstrict-aliasing and
        -Wconversion. It then proceeds to fix the few type conversion
        warnings that existed, casting explicitly where it's
        intentional, and replacing the data type where it's not.

        The -O0 and -Os flags are now added to CFLAGS depending on the
        rule invoked, also.

+++ boot/stage2/mmap.c
@@ -51,7 +51,7 @@ extern struct e820_map __mmap_new_map[MMAP_MAX_ENTRIES];
 static struct e820_map *const old_map = __mmap_old_map;
 static struct e820_map *const new_map = __mmap_new_map;
 
-static int mmap_clobber(struct e820_info *info)
+static size_t mmap_clobber(struct e820_info *info)
 {
     struct e820_map *mmap;
     size_t           nr_entries;
